% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gado_prop_gev11-ver0-1-2.R
\name{gado.prop_11}
\alias{gado.prop_11}
\title{Estimate the nonstationary GEV parameters}
\usage{
gado.prop_11(xdat, ntry = 20, ftol = 1e-06)
}
\arguments{
\item{xdat}{A numeric vector of data to be fitted. This data typically represents annual maximum values
of environmental variables like precipitation or river flow.}

\item{ntry}{The number of attempts for the optimization routine to avoid local minima and ensure a more
reliable estimation. The default is set to 20.}

\item{ftol}{The tolerance level for the optimization function, controlling the precision of the estimation.
The default value is set to 1e-6.}
}
\value{
A list containing the following elements:
\itemize{
 \item \code{para.prop}: Parameters estimated by the proposed method. For more details on the proposed methods, refer to
the paper "L-moment-based algorithm for nonstationary generalized extreme value model (2024)".
 \item \code{para.gado}: Parameters estimated by the GN16 method.
 \item \code{strup.sta}: Stationary parameters estimated by the Strup WLS method.
 \item \code{strup.org}: Original nonstationary parameters estimated by the Strup WLS method.
 \item \code{strup.final}: Final nonstationary parameters adjusted by the specified WLS method.
 \item \code{lme.sta}: Stationary parameters estimated by the L-moments method.
}
Each of these elements is a named vector of the GEV parameters, providing a comprehensive set of estimations
for further analysis.
}
\description{
This function estimates the parameters of a nonstationary Generalized Extreme Value (GEV) distribution
for a given dataset. It applies a combination of techniques including Weighted Least Squares (WLS),
robust regression, and the GN16 method, to handle both stationary and nonstationary data. The function
is designed to provide a robust estimation of GEV parameters, particularly useful in fields like hydrology
and climatology where extreme value analysis is critical. For more details on the methods mentioned above, refer to
the paper "L-moment-based algorithm for nonstationary generalized extreme value model (2024)".
}
\examples{
data("Trehafod")
result2 <- gado.prop_11(Trehafod$r1)
print(result2)

}
\author{
Jeong-Soo Park
}
