% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gado_prop_gev11-ver0-1-2.R
\name{nsgev}
\alias{nsgev}
\title{Estimate the nonstationary GEV parameters}
\usage{
nsgev(xdat, ntry = 20, ftol = 1e-06)
}
\arguments{
\item{xdat}{A numeric vector of data to be fitted. This is typically a series of annual maximums,
such as peak river flows or maximum daily temperatures.}

\item{ntry}{The number of attempts to perform in the optimization process. This parameter allows
the function to explore multiple starting points to avoid local minima and find a more global solution.
The default value is 20.}

\item{ftol}{The function tolerance for the optimization process. It determines the precision of the
optimization algorithm. A smaller value results in a more precise but potentially more time-consuming
optimization process. The default value is 1e-6.}
}
\value{
This function returns a list containing the following elements:
\itemize{
  \item \code{para.prop}: The proposed parameters for the nonstationary GEV model, including location,
  scale, and shape parameters, along with any nonstationarity coefficients.
  \item \code{precis}: The precision of the optimization process, indicating how closely the optimization
  criteria were met. A smaller value indicates a better fit.
}
}
\description{
This method combines L-moments, a goodness-of-fit measure, and robust regression,
using standardized residuals obtained through a transformation to a stationary sequence.
The function fits a nonstationary Generalized Extreme Value (GEV) distribution to the given data,
which is particularly useful in hydrology and climatology for modeling extreme events.
For more details, refer to the paper "L-moment-based algorithm for nonstationary generalized extreme value model (2024)".
}
\examples{
data("Trehafod")
result1 <- nsgev(Trehafod$r1)
print(result1)

}
\author{
Jeong-Soo Park
}
